requirements.txt
asgiref==3.11.0
Django==5.2.3
djangorestframework==3.16.1
psycopg2-binary==2.9.11
python-dotenv==1.2.1
sqlparse==0.5.4
pillow==10.0.0
djangorestframework_simplejwt==5.5.1